Add files via upload
Subroutines and Parameters

Purpose: This assignment is designed to give you practice with passing parameters to and returning values from ARM subroutines, and writing subroutine code that is independent of the calling program.

Problem description: Conventional computers require a multidimensional array to be stored as a linear sequence of elements. The compiler must, therefore, contain a procedure that computes the actual location of an element in an array from its specifications in terms of indices. Using the expression for computing the address of the a[i][j] element discussed in class, you are to write a subroutine in ARM that will receive three parameters: i, j, and n, where i is the row number, j is the column number, and n is the number of columns per row. The subroutine then computes and returns the offset that must be added to the starting address of the array to compute the address of a[i][j]. Your subroutine is computing only the offset, not the address. You are to name your subroutine offset.

For this problem, you may assume that i and j will be valid indices of a two-dimensional array, and n will be a valid value for the number of columns per row.

You are to also write a driver routine named main, that tests your offset function.

The user will enter i, j, and n as command-line arguments.